Coding style fixes

+++ tests/test_config.py
@@ -4,24 +4,25 @@ import json
 import mock
 import os
 from ..src import config
-from ..src.config import getOAuth
+from ..src.config import get_oauth
 
 
 testFile = 'test_oauth.json'
 
-def test_getOAuth_prompt(mocker):
-    os.remove(testFile)
+def test_get_oauth_prompt(mocker):
+    if os.path.exists(testFile):
+        os.remove(testFile)
     with mock.patch(config.__name__ + '.read', side_effect=['my_token', 'my_secret']):
-        assert getOAuth(testFile) == {
+        assert get_oauth(testFile) == {
             'consumer_token': 'my_token',
             'consumer_secret': 'my_secret'
         }
 
-def test_getOAuth_read(mocker):
+def test_get_oauth_read(mocker):
     auth = {
         'consumer_token': 'my_token',
         'consumer_secret': 'my_secret'
     }
     with open(testFile, 'w') as file:
         json.dump(auth, file)
-    assert getOAuth(testFile) == auth
+    assert get_oauth(testFile) == auth
